TG-211 Update form editor
Rebuild the form editor in a new more modern design. The new design features side buttons to insert new form elements.
Apart from that, the add and edit dialog for items was rebuild using a single dialog allowing to choose the item type.
The whole editor is in place and does not persist any changes to the server unless the changes are saved explicitly, since this can cause data loss the new editor saves its state in the IndexedDB of the browser